backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.8.3
supabase==2.9.1
Pillow
slowapi==0.1.9